refactor: split componts into separeted files
