+++ src/components/FeatureItem/styles.module.css
@@ -1,0 +1,4 @@
+.featureSvg {
+  height: 200px;
+  width: 200px;
+}
