+++ src/components/HomepageFeatures/styles.module.css
@@ -4,8 +4,3 @@
   padding: 2rem 0;
   width: 100%;
 }
-
-.featureSvg {
-  height: 200px;
-  width: 200px;
-}
