+++ src/components/HomepageHeader/styles.module.css
@@ -1,0 +1,37 @@
+
+.heroBanner {
+  padding: 4rem 0;
+  text-align: center;
+  position: relative;
+  overflow: hidden;
+  display: block;
+}
+
+.linkButtonCommunity {
+  background: linear-gradient(to left, #743ad5, #EC6737);
+  border: 0;
+  boxShadow: 6px 6px 26px -6px rgba(0,0,0,0.75);
+  transition: background 0.5s, color 0.5s, transform 0.5s;
+  transform: scale(1);
+}
+
+.linkButtonCommunity:hover{
+  background: linear-gradient(to left, #EC6737, #743ad5);
+  transform: scale(1.1);
+}
+
+@media screen and (max-width: 996px) {
+  .heroBanner {
+    padding: 2rem;
+  }
+
+  .linkButtonCommunity {
+    font-size: 0.8rem;
+    padding: 8px 16px;
+    width: 100%;
+  }
+  
+  .linkButtonCommunity:hover{
+    transform: none;
+  }
+}
